Generalize DraggableGizmo
