+++ src/Gizmos/axisDragGizmo.ts
@@ -1,43 +1,19 @@
-import { PointerDragBehavior } from "../Behaviors/Meshes/pointerDragBehavior";
-import { PointerInfo } from "../Events/pointerEvents";
 import { Color3 } from "../Maths/math.color";
 import { Matrix, Vector3 } from "../Maths/math.vector";
-import { AbstractMesh } from "../Meshes/abstractMesh";
 import { CylinderBuilder } from "../Meshes/Builders/cylinderBuilder";
 import { Mesh } from "../Meshes/mesh";
 import { TransformNode } from "../Meshes/transformNode";
-import { Observable, Observer } from "../Misc/observable";
 import { UtilityLayerRenderer } from "../Rendering/utilityLayerRenderer";
 import { Scene } from "../scene";
 import { Nullable } from "../types";
-import { Gizmo } from "./gizmo";
-import { GizmoMaterialSwitcher } from "./gizmoMaterialSwitcher";
+import { DraggableGizmo } from "./draggableGizmo";
 import { PositionGizmo } from "./positionGizmo";
 import { StandardMaterial } from "../Materials/standardMaterial";
 /**
  * Single axis drag gizmo
  */
-export class AxisDragGizmo extends Gizmo {
-    /**
-     * Drag behavior responsible for the gizmos dragging interactions
-     */
-    public dragBehavior: PointerDragBehavior;
-    private _pointerObserver: Nullable<Observer<PointerInfo>> = null;
-    /**
-     * Drag distance in babylon units that the gizmo will snap to when dragged (Default: 0)
-     */
-    public snapDistance = 0;
-    /**
-     * Event that fires each time the gizmo snaps to a new location.
-     * * snapDistance is the the change in distance
-     */
-    public onSnapObservable = new Observable<{ snapDistance: number }>();
-
-    private _isEnabled: boolean = true;
-    private _parent: Nullable<PositionGizmo> = null;
-
+export class AxisDragGizmo extends DraggableGizmo {
     private _arrow: TransformNode;
-    private _materialSwitcher: GizmoMaterialSwitcher;
 
     /** @hidden */
     public static _CreateArrow(scene: Scene, material?: StandardMaterial): TransformNode {
@@ -73,8 +49,7 @@ export class AxisDragGizmo extends Gizmo {
      * @param color The color of the gizmo
      */
     constructor(dragAxis: Vector3, color: Color3 = Color3.Gray(), gizmoLayer: UtilityLayerRenderer = UtilityLayerRenderer.DefaultUtilityLayer, parent: Nullable<PositionGizmo> = null) {
-        super(gizmoLayer);
-        this._parent = parent;
+        super({ dragAxis }, color, gizmoLayer, parent);
 
         // Build mesh on root node
         this._arrow = AxisDragGizmo._CreateArrow(gizmoLayer.utilityLayerScene);
@@ -83,18 +58,6 @@ export class AxisDragGizmo extends Gizmo {
         this._arrow.scaling.scaleInPlace(1 / 3);
         this._arrow.parent = this._rootMesh;
 
-        // Add drag behavior to handle events when the gizmo is dragged
-        this.dragBehavior = new PointerDragBehavior({ dragAxis: dragAxis });
-        this.dragBehavior.moveAttached = false;
-        this._rootMesh.addBehavior(this.dragBehavior);
-
-        // Create Material Switcher
-        this._materialSwitcher = new GizmoMaterialSwitcher(
-            color,
-            this.dragBehavior,
-            gizmoLayer._getSharedGizmoLight(),
-            gizmoLayer.utilityLayerScene
-        );
         this._materialSwitcher.registerMeshes(
             this._rootMesh.getChildMeshes(false)
         );
@@ -132,62 +95,5 @@ export class AxisDragGizmo extends Gizmo {
                 }
             }
         });
-    }
-    protected _attachedMeshChanged(value: Nullable<AbstractMesh>) {
-        if (this.dragBehavior) {
-            this.dragBehavior.enabled = value ? true : false;
-        }
-    }
-
-    /**
-     * If the gizmo is enabled
-     */
-    public set isEnabled(value: boolean) {
-        this._isEnabled = value;
-        if (!value) {
-            this.attachedMesh = null;
-        }
-        else {
-            if (this._parent) {
-                this.attachedMesh = this._parent.attachedMesh;
-            }
-        }
-    }
-    public get isEnabled(): boolean {
-        return this._isEnabled;
-    }
-
-    public get materialSwitcher() {
-        return this._materialSwitcher;
-    }
-
-    /**
-     * Disposes of the gizmo
-     */
-    public dispose() {
-        this.onSnapObservable.clear();
-        this.gizmoLayer.utilityLayerScene.onPointerObservable.remove(this._pointerObserver);
-        this.dragBehavior.detach();
-        this._materialSwitcher.dispose();
-        super.dispose();
-    }
-
-    /**
-     * Disposes and replaces the current meshes in the gizmo with the specified mesh
-     * @param mesh The mesh to replace the default mesh of the gizmo
-     * @param useGizmoMaterials If the gizmo's default materials should be used (default: false)
-     */
-    public setCustomMesh(mesh: Mesh, useGizmoMaterials: boolean = false) {
-        this._materialSwitcher.unregisterMeshes(
-            this._rootMesh.getChildMeshes()
-        );
-
-        super.setCustomMesh(mesh);
-
-        if (useGizmoMaterials) {
-            this._materialSwitcher.registerMeshes(
-                this._rootMesh.getChildMeshes()
-            );
-        }
     }
 }
